// src/app/core/models/ro-le.ts
export class RoLeList {
  MAPMIS: string = '';
  TENCONGTY: string = '';
  TRUYENTAIDIEN: string = '';
  TENROLE: string = '';
  TENTRAM: string = '';
}

export class RoLeDetail {
  Mapmis: string | null | undefined;
  Madvql: string | null | undefined;
  Tencongty: string | null | undefined;
  Truyentaidien: string | null | undefined;
  Tenrole: string | null | undefined;
  Sohieu: string | null | undefined;
  Sohuu: string | null | undefined;
  Ngaylapdat: Date | null | undefined;
  Ngayvh: Date | null | undefined;
  Thuoctram: string | null | undefined;
  Tentram: string | null | undefined;
  Hangsx: string | null | undefined;
  Soserial: string | null | undefined;
  Sohieubanve: string | null | undefined;
  Sododanhso: string | null | undefined;
  Mach: string | null | undefined;
  Tbbaove: string | null | undefined;
  Tubv: string | null | undefined;
  Cottenhienthi: string | null | undefined;
  Dahienthitrensd: string | null | undefined;
  Hienthiten: string | null | undefined;
  Hoatdong: string | null | undefined;
  Tthientai: string | null | undefined;
  Jsongeo: string | null | undefined;
  Maudong: string | null | undefined;
  Maucat: string | null | undefined;
  Daunoidau: string | null | undefined;
  Daunoicuoi: string | null | undefined;
  Ghichu: string | null | undefined;
}